Backend_TalentaTalk/TalentaTalkBackend/requirements.txt
fastapi
uvicorn
starlette
python-multipart
sqlalchemy[asyncio]
psycopg2-binary
python-dotenv
passlib[bcrypt]
bcrypt
python-jose[cryptography]
pydantic[email]
email-validator
itsdangerous
aiohttp
httpx
numpy
pandas
librosa
torch
transformers
openai-whisper
gruut
google-generativeai
openpyxl
xlsxwriter
pandas
ffmpeg-python
soundfile